moved ioi_cace from global to local to each client
Originally I had left this as a global map since the same tag name should yield the same ioi even on different clients, but multiple clients using the same map will panic if you read a new tag (so we're adding to the cache) at the same time another client reads an existing tag.  I could have put an rwmutex on it, but moving it to client scope was simple enough.

Frankly, we probably dont' even need the ioi cache.  It's not that computationally expensive to calculate them every time instead of looking them up.