Better unique number generation.  Multiple client support.  Auto Reconnect support.
There are a number of things in the protocol that need unique numbers to function.  Previously some of these were hardcoded.  Now they are coming from a sequencer.  The sequencer seeds with a random value in the init function.

Presumably the sequencer could collide with another instance running, but it's probably rare enough to be OK.